Gamma, third update (13th commit)
Mostly SEO upgrades, alt text, titles, and meta stuff. Condensed the
keywords down to five important ones because that supposedly works
better, helps it focus I guess. Added CTA navigation to the bottom of
each project and logically hide one or the other if its at the end of
the project list. There's a new loop in the project.php page that finds
the total number of items in a DB table. I may want to move that to a
class if I end up using it a lot.